requirements.txt
python-dotenv==1.0.0
pyrogram==2.0.106
tgcrypto==1.2.5
pytgcalls==3.0.0.dev24
yt-dlp==2023.11.16
PyTgCalls==3.0.0.dev24
python-telegram-bot==20.7

